Add Doublet Detection page implementation
This implements the 'Scrublet' doublet detection algorithm from scanpy.
Changes were also made to quality_control.py to incorporate a button
that triggers the filtering.  This is to allow a user to revist a past
page and return to a later page without any issues.  There are not yet
any sophisticated controls to reset later stages of the workflow based
on changes the user has made to selections.