Add text explaining analysis to last four pages
This update is to make it easier for the user to understand the
rationale behind the step. Other changes to improve readability were
also made. This includes adding a warning icon to text that requires
the user's attention - such as when a step has not yet been run. The
placeholder "More Information" textbox expanders that were added for
all plots were commented out if the placeholder text had not been
replaced.